Reworked the item class structure.
