+++ test/unit/inventory_test.cpp
@@ -1,8 +1,8 @@
 #include <string>
 #include <catch.hpp>
-#include <inventory.cpp>
+#include <item.h>
 
-SCENARIO(" All Items have a name, ItemProperty, count, and durability.", "[Item]")
+SCENARIO(" All Items have a proper name, ItemProperty, count, and durability.", "[Item]")
 {
   GIVEN("An Item with 'Sword' as its name.")
   {
@@ -11,7 +11,8 @@ SCENARIO(" All Items have a name, ItemProperty, count, and durability.", "[Item]
 
     REQUIRE(item.getCount() > 0);
 
-    REQUIRE(item.getDurability() >= 0);
+    REQUIRE(0 <= item.getDurability());
+    REQUIRE(item.getDurability() <= 100);
 
 
     ItemProperty prop = item.getItemProperty();
@@ -22,16 +23,3 @@ SCENARIO(" All Items have a name, ItemProperty, count, and durability.", "[Item]
   }
 }
 
-
-SCENARIO(" All Wearables have a defense stat bigger than or equal to 0. " )
-{
-  GIVEN("　A Wearable named 'T-Shirt'. ")
-  {
-    Armor shirt("T-Shirt");
-    REQUIRE(shirt.getDefense() >= 0);
-
-    Armor sweater("Turtleneck", 1, 100, -20);
-    REQUIRE(sweater.getDefense() >= 0);
-
-  }
-}
